+++ src/service.h
@@ -3,23 +3,10 @@
 
 #include <string>
 
+#include "config.h"
+
 namespace csi {
 namespace service {
-class Config {
- public:
-  Config(const std::string endpoint, const std::string driver_name,
-         const std::string version);
-  ~Config();
-
-  const std::string &driver_name() const { return driver_name_; }
-  const std::string &endpoint() const { return endpoint_; }
-  const std::string &version() const { return version_; }
-
- private:
-  std::string driver_name_;
-  std::string endpoint_;
-  std::string version_;
-};
 class Server {
  public:
   Server(Config config);
